+++ dao/src/main/java/com/epam/mediaserver/dao/impl/SqlOrderSongDao.java
@@ -78,8 +78,8 @@ public class SqlOrderSongDao extends AbstractModelDao<OrderSong> implements Orde
         PreparedStatement ps = con.prepareStatement(query);
 
         OrderSong orderSong = (OrderSong) model;
-        ps.setInt(1, orderSong.getOrder().getId());
-        ps.setInt(2, orderSong.getSong().getId());
+        ps.setLong(1, orderSong.getOrder().getId());
+        ps.setLong(2, orderSong.getSong().getId());
 
         return ps.executeUpdate();
     }
@@ -89,9 +89,9 @@ public class SqlOrderSongDao extends AbstractModelDao<OrderSong> implements Orde
         PreparedStatement ps = con.prepareStatement(query);
 
         OrderSong orderSong = (OrderSong) model;
-        ps.setInt(1, orderSong.getSong().getId());
-        ps.setInt(2, orderSong.getOrder().getId());
-        ps.setInt(3, orderSong.getId());
+        ps.setLong(1, orderSong.getSong().getId());
+        ps.setLong(2, orderSong.getOrder().getId());
+        ps.setLong(3, orderSong.getId());
 
         return ps.executeUpdate();
     }
@@ -101,7 +101,7 @@ public class SqlOrderSongDao extends AbstractModelDao<OrderSong> implements Orde
         PreparedStatement ps = con.prepareStatement(query);
         OrderSong orderSong = (OrderSong) model;
 
-        ps.setInt(1, orderSong.getId());
+        ps.setLong(1, orderSong.getId());
 
         return ps.executeUpdate();
     }
@@ -137,9 +137,9 @@ public class SqlOrderSongDao extends AbstractModelDao<OrderSong> implements Orde
         OrderSong orderSong = new OrderSong();
 
         try {
-            orderSong.setId(rs.getInt(ORDER_SONG_ID));
-            orderSong.setSong((Song) songDao.getById(rs.getInt(SONG_ID)));
-            orderSong.setOrder((Order) orderDao.getById(rs.getInt(ORDER_ID)));
+            orderSong.setId(rs.getLong(ORDER_SONG_ID));
+            orderSong.setSong((Song) songDao.getById(   rs.getLong(SONG_ID)));
+            orderSong.setOrder((Order) orderDao.getById(rs.getLong(ORDER_ID)));
 
         } catch (SQLException e) {
             LOGGER.error("SQL Exception");
