+++ service/src/main/java/com/epam/mediaserver/service/impl/BonusTableService.java
@@ -54,7 +54,7 @@ public class BonusTableService implements BonusService{
 
     }
 
-    public void edit(int id, String title, String description, String discount, String code)
+    public void edit(Long id, String title, String description, String discount, String code)
         throws ServiceException, ValidateException {
 
         try {
@@ -80,7 +80,7 @@ public class BonusTableService implements BonusService{
         }
     }
 
-    public void delete(int id) throws ServiceException {
+    public void delete(Long id) throws ServiceException {
 
         try {
             Bonus bonus = (Bonus) bonusDao.getById(id);
