+++ service/src/main/java/com/epam/mediaserver/service/impl/GenreTableService.java
@@ -70,7 +70,7 @@ public class GenreTableService {
 
     }
 
-    public void edit(int id, String title, String description, String image)
+    public void edit(Long id, String title, String description, String image)
         throws ServiceException, ValidateException {
 
         try {
@@ -96,7 +96,7 @@ public class GenreTableService {
         }
     }
 
-    public void delete(int id) throws ServiceException {
+    public void delete(Long id) throws ServiceException {
 
         try {
             Genre genre = genreDao.getById(id);
